fix(test): test local config last, since it has side effects
